+++ src/roq/deribit/web_socket.cpp
@@ -476,5 +476,9 @@ void WebSocket::operator()(const server::Trace<json::Changes> &) {
   log::fatal("Unexpected"_sv);
 }
 
+void WebSocket::operator()(const server::Trace<json::Order> &) {
+  log::fatal("Unexpected"_sv);
+}
+
 }  // namespace deribit
 }  // namespace roq